[refactor] Apply Interface to User service
